separated DASH into its own file going to bed
